requirements.txt
astor==0.8.1
attrs==21.4.0
Babel==2.9.1
bce-python-sdk==0.8.64
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
cycler==0.11.0
Cython==0.29.26
cython-bbox==0.1.3
decorator==5.1.1
distlib==0.3.4
dominate==2.6.0
et-xmlfile==1.1.0
filelock==3.4.2
flake8==4.0.1
flake8-import-order==0.18.1
Flask==2.0.2
Flask-Babel==2.0.0
fonttools==4.28.5
future==0.18.2
GPUtil==1.4.0
identify==2.4.4
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
lap==0.4.0
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
motmetrics==1.2.0
nodeenv==1.6.0
numpy
olefile
opencv-python==4.5.5.62
openpyxl==3.0.9
packaging==21.3
paddledet==2.3.0
paddlepaddle-gpu
pandas==1.3.5
Pillow==8.4.0
pip-review==1.1.1
platformdirs==2.4.1
pluggy==1.0.0
pre-commit==2.17.0
protobuf==3.19.3
py==1.11.0
py-cpuinfo==8.0.0
pycocotools==2.0.3
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.12.0
pyflakes==2.4.0
pyparsing==3.0.6
pytest==6.2.5
pytest-benchmark==3.4.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
resolvelib==0.5.4
scikit-learn==1.0.2
scipy==1.7.3
Shapely==1.8.0
shellcheck-py==0.8.0.3
six
sklearn==0.0
terminaltables==3.1.10
threadpoolctl==3.0.0
toml==0.10.2
tqdm==4.62.3
typeguard==2.13.3
urllib3==1.26.8
virtualenv==20.13.0
visualdl==2.2.3
Werkzeug==2.0.2
xmltodict==0.12.0
